bugfix: harden heterogeneous pd cache restoration.
- serialize the complete heterogeneous cache restore transaction\n- keep model-specific resharding in SpecKVCacheTransfer\n- derive staging layout from runtime config and expose parallel pull via DisaggPDConfig
