+++ xllm/core/framework/kv_cache_transfer/spec_kv_cache_transfer.cpp
@@ -17,14 +17,24 @@ limitations under the License.
 
 #include <glog/logging.h>
 
+#include <algorithm>
+#include <exception>
+#include <functional>
+#include <numeric>
 #include <optional>
 
 #include "common/macros.h"
+#include "core/framework/config/disagg_pd_config.h"
+#include "core/framework/config/kv_cache_config.h"
+#include "core/framework/config/scheduler_config.h"
 #include "util/timer.h"
 
 namespace xllm {
 
 namespace {
+
+constexpr int64_t kSupportedHeterogeneousSourceShardCount = 2;
+constexpr int64_t kKeyValueShardDimension = 2;
 
 std::optional<int32_t> get_remote_tp_size(
     const std::vector<TransferKVInfo>& transfer_kv_infos) {
@@ -87,6 +97,168 @@ void merge_heterogeneous_kv_blocks(
   }
 }
 
+bool is_linear_state_cache(KVCacheTensorRole role) {
+  return role == KVCacheTensorRole::CONV || role == KVCacheTensorRole::SSM;
+}
+
+int64_t sharded_dimension(KVCacheTensorRole role, const torch::Tensor& tensor) {
+  if (role == KVCacheTensorRole::KEY || role == KVCacheTensorRole::VALUE) {
+    return kKeyValueShardDimension;
+  }
+  if (role == KVCacheTensorRole::CONV) {
+    return tensor.dim() - 1;
+  }
+  if (role == KVCacheTensorRole::SSM) {
+    return 1;
+  }
+  return -1;
+}
+
+int64_t get_checkpoint_stride(
+    const std::vector<RegisteredCache>& layer_caches) {
+  const RegisteredCache* conv_cache = nullptr;
+  const RegisteredCache* ssm_cache = nullptr;
+  for (const RegisteredCache& cache : layer_caches) {
+    if (!cache.tensor.defined()) {
+      continue;
+    }
+    if (cache.role == KVCacheTensorRole::CONV) {
+      conv_cache = &cache;
+    } else if (cache.role == KVCacheTensorRole::SSM) {
+      ssm_cache = &cache;
+    }
+  }
+  if (conv_cache == nullptr || ssm_cache == nullptr) {
+    return 1;
+  }
+  CHECK_GT(conv_cache->tensor.size(0), 0);
+  CHECK_EQ(ssm_cache->tensor.size(0) % conv_cache->tensor.size(0), 0);
+  return ssm_cache->tensor.size(0) / conv_cache->tensor.size(0);
+}
+
+std::vector<uint64_t> make_compact_ids(size_t count) {
+  std::vector<uint64_t> ids(count);
+  std::iota(ids.begin(), ids.end(), 0);
+  return ids;
+}
+
+std::vector<uint64_t> expand_checkpoint_ids(
+    const std::vector<uint64_t>& logical_ids,
+    int64_t checkpoint_stride) {
+  std::vector<uint64_t> ids;
+  ids.reserve(logical_ids.size() * static_cast<size_t>(checkpoint_stride));
+  for (uint64_t logical_id : logical_ids) {
+    for (int64_t checkpoint = 0; checkpoint < checkpoint_stride; ++checkpoint) {
+      ids.push_back(logical_id * static_cast<uint64_t>(checkpoint_stride) +
+                    static_cast<uint64_t>(checkpoint));
+    }
+  }
+  return ids;
+}
+
+std::vector<torch::Tensor> make_staging_shard_views(
+    const torch::Tensor& staging_tensor,
+    int64_t source_shard_count,
+    int64_t row_count) {
+  CHECK_GT(source_shard_count, 0);
+  CHECK_EQ(staging_tensor.size(0) % source_shard_count, 0);
+  const int64_t rows_per_shard = staging_tensor.size(0) / source_shard_count;
+  CHECK_LE(row_count, rows_per_shard);
+  std::vector<torch::Tensor> shard_tensors;
+  shard_tensors.reserve(static_cast<size_t>(source_shard_count));
+  for (int64_t shard = 0; shard < source_shard_count; ++shard) {
+    shard_tensors.emplace_back(
+        staging_tensor.narrow(0, shard * rows_per_shard, row_count));
+  }
+  return shard_tensors;
+}
+
+torch::Tensor merge_shard_tensors(
+    KVCacheTensorRole role,
+    const torch::Tensor& staging_tensor,
+    const std::vector<RegisteredCache>& layer_staging_caches,
+    const std::vector<torch::Tensor>& shard_tensors,
+    int64_t shard_dim,
+    int64_t layer_id) {
+  if (role != KVCacheTensorRole::CONV) {
+    return torch::cat(shard_tensors, shard_dim);
+  }
+
+  int64_t local_v_width = -1;
+  for (const RegisteredCache& candidate : layer_staging_caches) {
+    if (candidate.role == KVCacheTensorRole::SSM &&
+        candidate.tensor.defined()) {
+      local_v_width = candidate.tensor.size(1) * candidate.tensor.size(3);
+      break;
+    }
+  }
+  CHECK_GT(local_v_width, 0)
+      << "CONV cache requires a matching SSM cache at layer " << layer_id;
+  const int64_t local_conv_width = staging_tensor.size(shard_dim);
+  CHECK_EQ((local_conv_width - local_v_width) % 2, 0)
+      << "Invalid Q/K/V CONV cache layout at layer " << layer_id;
+  const int64_t local_qk_width = (local_conv_width - local_v_width) / 2;
+  std::vector<torch::Tensor> q_shards;
+  std::vector<torch::Tensor> k_shards;
+  std::vector<torch::Tensor> v_shards;
+  q_shards.reserve(shard_tensors.size());
+  k_shards.reserve(shard_tensors.size());
+  v_shards.reserve(shard_tensors.size());
+  for (const torch::Tensor& shard : shard_tensors) {
+    std::vector<torch::Tensor> qkv = torch::split_with_sizes(
+        shard, {local_qk_width, local_qk_width, local_v_width}, shard_dim);
+    q_shards.emplace_back(qkv[0]);
+    k_shards.emplace_back(qkv[1]);
+    v_shards.emplace_back(qkv[2]);
+  }
+  return torch::cat({torch::cat(q_shards, shard_dim),
+                     torch::cat(k_shards, shard_dim),
+                     torch::cat(v_shards, shard_dim)},
+                    shard_dim);
+}
+
+void copy_merged_cache_rows(const RegisteredCache& registered_cache,
+                            const std::vector<uint64_t>& final_ids,
+                            const torch::Tensor& merged) {
+  std::vector<int64_t> signed_final_ids(final_ids.begin(), final_ids.end());
+  torch::Tensor final_indices =
+      torch::tensor(signed_final_ids,
+                    torch::TensorOptions().dtype(torch::kLong))
+          .to(registered_cache.tensor.device(), /*non_blocking=*/false);
+  registered_cache.tensor.index_copy_(0, final_indices, merged);
+}
+
+torch::Tensor make_page_aligned_staging_tensor(
+    std::vector<int64_t>& shape,
+    const torch::TensorOptions& options,
+    int64_t element_size) {
+  constexpr int64_t kHcclPageSize = 2 * 1024 * 1024;
+  int64_t row_elements = 1;
+  for (size_t dim = 1; dim < shape.size(); ++dim) {
+    row_elements *= shape[dim];
+  }
+  const int64_t row_bytes = row_elements * element_size;
+  const int64_t rows_per_alignment =
+      kHcclPageSize / std::gcd(kHcclPageSize, row_bytes);
+  shape[0] = ((shape[0] + rows_per_alignment - 1) / rows_per_alignment) *
+             rows_per_alignment;
+
+  const int64_t aligned_numel = std::accumulate(
+      shape.begin(), shape.end(), int64_t{1}, std::multiplies<int64_t>());
+  const int64_t padding_elements = kHcclPageSize / element_size;
+  torch::Tensor backing =
+      torch::empty({aligned_numel + padding_elements}, options);
+  const uintptr_t base = reinterpret_cast<uintptr_t>(backing.data_ptr());
+  const uintptr_t aligned = (base + kHcclPageSize - 1) & ~(kHcclPageSize - 1);
+  const int64_t offset_elements =
+      static_cast<int64_t>(aligned - base) / element_size;
+  torch::Tensor stage =
+      backing.narrow(0, offset_elements, aligned_numel).view(shape);
+  CHECK_EQ(reinterpret_cast<uintptr_t>(stage.data_ptr()) % kHcclPageSize, 0);
+  CHECK_EQ(stage.numel() * element_size % kHcclPageSize, 0);
+  return stage;
+}
+
 }  // namespace
 
 SpecKVCacheTransfer::SpecKVCacheTransfer(const uint16_t listen_port,
@@ -97,6 +269,8 @@ SpecKVCacheTransfer::SpecKVCacheTransfer(const uint16_t listen_port,
     : LlmDataDistTransfer(listen_port, instance_role, enable_lighting_indexer) {
   enable_mla_ = enable_mla;
   draft_body_uses_tp1_ = draft_body_uses_tp1;
+  parallel_shard_pull_ =
+      DisaggPDConfig::get_instance().enable_pd_parallel_shard_pull();
 }
 
 void SpecKVCacheTransfer::register_kv_cache(
@@ -121,12 +295,444 @@ void SpecKVCacheTransfer::register_kv_cache_spec(
   const bool source_is_sharded = role_ == LlmRole::kPrompt;
   register_hetero_staging_caches(layer_registered_caches_,
                                  hetero_staging_registered_caches_,
-                                 /*source_shard_count=*/2,
+                                 kSupportedHeterogeneousSourceShardCount,
                                  source_is_sharded);
   register_hetero_staging_caches(spec_layer_registered_caches_,
                                  spec_hetero_staging_registered_caches_,
-                                 /*source_shard_count=*/2,
+                                 kSupportedHeterogeneousSourceShardCount,
                                  source_is_sharded && !draft_body_uses_tp1_);
+}
+
+bool SpecKVCacheTransfer::pull_and_merge_sharded_caches(
+    const LayerRegisteredCaches& layer_registered_caches,
+    const LayerRegisteredCaches& staging_registered_caches,
+    const std::vector<uint64_t>& src_cluster_ids,
+    const std::vector<uint64_t>& src_blocks,
+    const std::vector<uint64_t>& dst_blocks,
+    const std::vector<uint64_t>& src_linear_state_ids,
+    const std::vector<uint64_t>& dst_linear_state_ids) {
+  if (src_cluster_ids.size() !=
+          static_cast<size_t>(kSupportedHeterogeneousSourceShardCount) ||
+      layer_registered_caches.size() != staging_registered_caches.size() ||
+      src_blocks.size() != dst_blocks.size() ||
+      src_linear_state_ids.size() != dst_linear_state_ids.size()) {
+    LOG(ERROR) << "Invalid heterogeneous KV pull metadata: src_tp="
+               << src_cluster_ids.size() << ", src_blocks=" << src_blocks.size()
+               << ", dst_blocks=" << dst_blocks.size()
+               << ", src_linear_states=" << src_linear_state_ids.size()
+               << ", dst_linear_states=" << dst_linear_state_ids.size();
+    return false;
+  }
+
+  const int64_t shard_count = static_cast<int64_t>(src_cluster_ids.size());
+  Timer breakdown_total_timer;
+  double pull_seconds = 0.0;
+  double pull_wall_seconds = 0.0;
+  double merge_seconds = 0.0;
+  double conv_pull_seconds = 0.0;
+  double conv_merge_seconds = 0.0;
+  double ssm_pull_seconds = 0.0;
+  double ssm_merge_seconds = 0.0;
+  std::vector<double> shard_pull_seconds(src_cluster_ids.size(), 0.0);
+  size_t pull_calls = 0;
+  size_t merge_calls = 0;
+  bool success = true;
+  for (int64_t layer_id = 0;
+       layer_id < static_cast<int64_t>(layer_registered_caches.size());
+       ++layer_id) {
+    const auto& layer_caches = layer_registered_caches[layer_id];
+    const auto& layer_staging_caches = staging_registered_caches[layer_id];
+    if (layer_caches.size() != layer_staging_caches.size()) {
+      LOG(ERROR) << "Heterogeneous KV staging layout mismatch at layer "
+                 << layer_id;
+      return false;
+    }
+    const int64_t checkpoint_stride = get_checkpoint_stride(layer_caches);
+
+    for (size_t cache_index = 0; cache_index < layer_caches.size();
+         ++cache_index) {
+      const RegisteredCache& registered_cache = layer_caches[cache_index];
+      const RegisteredCache& stage_cache = layer_staging_caches[cache_index];
+      const int64_t shard_dim =
+          sharded_dimension(registered_cache.role, registered_cache.tensor);
+      if (shard_dim < 0) {
+        LOG(ERROR) << "Unsupported heterogeneous KV tensor role: "
+                   << registered_cache.role.to_string();
+        return false;
+      }
+      CHECK_EQ(registered_cache.tensor.size(shard_dim) % shard_count, 0)
+          << "Cache shard dimension is not divisible by source shard count, "
+          << "layer=" << layer_id
+          << ", role=" << registered_cache.role.to_string()
+          << ", shape=" << registered_cache.tensor.sizes();
+
+      const bool linear_state_cache =
+          is_linear_state_cache(registered_cache.role);
+      std::vector<uint64_t> remote_ids =
+          linear_state_cache ? src_linear_state_ids : src_blocks;
+      std::vector<uint64_t> final_ids =
+          linear_state_cache ? dst_linear_state_ids : dst_blocks;
+      if (registered_cache.role == KVCacheTensorRole::SSM) {
+        remote_ids = expand_checkpoint_ids(remote_ids, checkpoint_stride);
+        final_ids = expand_checkpoint_ids(final_ids, checkpoint_stride);
+      }
+      if (remote_ids.empty()) {
+        continue;
+      }
+
+      if (stage_cache.tensor.size(0) <
+          static_cast<int64_t>(remote_ids.size())) {
+        LOG(ERROR) << "Heterogeneous KV transfer exceeds staging capacity, "
+                   << "layer=" << layer_id
+                   << ", role=" << registered_cache.role.to_string()
+                   << ", requested=" << remote_ids.size()
+                   << ", capacity=" << stage_cache.tensor.size(0);
+        return false;
+      }
+      CHECK_EQ(stage_cache.tensor.size(0) % shard_count, 0);
+      const int64_t rows_per_shard = stage_cache.tensor.size(0) / shard_count;
+      CHECK_LE(static_cast<int64_t>(remote_ids.size()), rows_per_shard);
+
+      std::vector<llm_datadist::Status> pull_rets(src_cluster_ids.size(),
+                                                  LLM_SUCCESS);
+      std::vector<double> current_pull_seconds(src_cluster_ids.size(), 0.0);
+      auto pull_one_shard = [&](size_t source_shard) {
+        const uint64_t src_cluster_id = src_cluster_ids[source_shard];
+        std::vector<uint64_t> staging_ids = make_compact_ids(remote_ids.size());
+        const uint64_t staging_offset =
+            static_cast<uint64_t>(source_shard * rows_per_shard);
+        for (uint64_t& staging_id : staging_ids) {
+          staging_id += staging_offset;
+        }
+        CacheIndex src_cache_index{src_cluster_id,
+                                   registered_cache.cache.cache_id};
+        KvCacheExtParam ext_param{};
+        ext_param.src_layer_range = {0, 0};
+        ext_param.dst_layer_range = {0, 0};
+        ext_param.tensor_num_per_layer = 1;
+        Timer pull_timer;
+        pull_rets[source_shard] =
+            llm_data_dist_->PullKvBlocks(src_cache_index,
+                                         stage_cache.cache,
+                                         remote_ids,
+                                         staging_ids,
+                                         ext_param);
+        current_pull_seconds[source_shard] = pull_timer.elapsed_seconds();
+      };
+
+      Timer pull_group_timer;
+      if (parallel_shard_pull_) {
+        TaskGroup shard_pull_group(1);
+        shard_pull_threadpool_.schedule(
+            shard_pull_group.wrap([&]() { pull_one_shard(1); }));
+        std::exception_ptr request_thread_exception;
+        try {
+          pull_one_shard(0);
+        } catch (...) {
+          request_thread_exception = std::current_exception();
+        }
+        shard_pull_group.wait();
+        if (request_thread_exception) {
+          std::rethrow_exception(request_thread_exception);
+        }
+      } else {
+        for (size_t source_shard = 0; source_shard < src_cluster_ids.size();
+             ++source_shard) {
+          pull_one_shard(source_shard);
+        }
+      }
+      pull_wall_seconds += pull_group_timer.elapsed_seconds();
+
+      for (size_t source_shard = 0; source_shard < src_cluster_ids.size();
+           ++source_shard) {
+        const uint64_t src_cluster_id = src_cluster_ids[source_shard];
+        pull_seconds += current_pull_seconds[source_shard];
+        shard_pull_seconds[source_shard] += current_pull_seconds[source_shard];
+        ++pull_calls;
+        if (registered_cache.role == KVCacheTensorRole::CONV) {
+          conv_pull_seconds += current_pull_seconds[source_shard];
+        } else if (registered_cache.role == KVCacheTensorRole::SSM) {
+          ssm_pull_seconds += current_pull_seconds[source_shard];
+        }
+        if (pull_rets[source_shard] != LLM_SUCCESS) {
+          LOG(ERROR) << "Heterogeneous PullKvBlocks failed, layer=" << layer_id
+                     << ", role=" << registered_cache.role.to_string()
+                     << ", src_cluster_id=" << src_cluster_id
+                     << ", src_cache_id=" << registered_cache.cache.cache_id
+                     << ", ret=" << std::hex << pull_rets[source_shard];
+          success = false;
+          break;
+        }
+      }
+
+      if (success) {
+        Timer merge_timer;
+        const std::vector<torch::Tensor> shard_tensors =
+            make_staging_shard_views(stage_cache.tensor,
+                                     shard_count,
+                                     static_cast<int64_t>(remote_ids.size()));
+        const torch::Tensor merged = merge_shard_tensors(registered_cache.role,
+                                                         stage_cache.tensor,
+                                                         layer_staging_caches,
+                                                         shard_tensors,
+                                                         shard_dim,
+                                                         layer_id);
+        copy_merged_cache_rows(registered_cache, final_ids, merged);
+        const double current_merge_seconds = merge_timer.elapsed_seconds();
+        merge_seconds += current_merge_seconds;
+        ++merge_calls;
+        if (registered_cache.role == KVCacheTensorRole::CONV) {
+          conv_merge_seconds += current_merge_seconds;
+        } else if (registered_cache.role == KVCacheTensorRole::SSM) {
+          ssm_merge_seconds += current_merge_seconds;
+        }
+      }
+      if (!success) {
+        return false;
+      }
+    }
+  }
+  LOG(INFO) << "[PD-PERF] Heterogeneous pull-merge breakdown"
+            << " linear_request=" << !src_linear_state_ids.empty()
+            << " parallel_shard_pull=" << parallel_shard_pull_
+            << " pull_calls=" << pull_calls << " merge_calls=" << merge_calls
+            << " shard0_pull_ms=" << shard_pull_seconds[0] * 1000.0
+            << " shard1_pull_ms=" << shard_pull_seconds[1] * 1000.0
+            << " conv_pull_ms=" << conv_pull_seconds * 1000.0
+            << " ssm_pull_ms=" << ssm_pull_seconds * 1000.0
+            << " pull_work_ms=" << pull_seconds * 1000.0
+            << " pull_ms=" << pull_wall_seconds * 1000.0
+            << " conv_merge_ms=" << conv_merge_seconds * 1000.0
+            << " ssm_merge_ms=" << ssm_merge_seconds * 1000.0
+            << " merge_ms=" << merge_seconds * 1000.0 << " other_ms="
+            << (breakdown_total_timer.elapsed_seconds() - pull_wall_seconds -
+                merge_seconds) *
+                   1000.0
+            << " total_ms=" << breakdown_total_timer.elapsed_seconds() * 1000.0;
+  return true;
+}
+
+bool SpecKVCacheTransfer::merge_pre_pushed_sharded_caches(
+    const LayerRegisteredCaches& layer_registered_caches,
+    const LayerRegisteredCaches& staging_registered_caches,
+    const std::vector<uint64_t>& dst_blocks,
+    const std::vector<uint64_t>& dst_linear_state_ids,
+    int64_t source_shard_count) {
+  if (source_shard_count != kSupportedHeterogeneousSourceShardCount ||
+      layer_registered_caches.size() != staging_registered_caches.size()) {
+    LOG(ERROR) << "Invalid pre-pushed heterogeneous KV layout: source_tp="
+               << source_shard_count;
+    return false;
+  }
+
+  for (size_t layer_id = 0; layer_id < layer_registered_caches.size();
+       ++layer_id) {
+    const auto& layer_caches = layer_registered_caches[layer_id];
+    const auto& layer_staging_caches = staging_registered_caches[layer_id];
+    if (layer_caches.size() != layer_staging_caches.size()) {
+      LOG(ERROR) << "Pre-pushed KV staging layout mismatch at layer "
+                 << layer_id;
+      return false;
+    }
+    const int64_t checkpoint_stride = get_checkpoint_stride(layer_caches);
+
+    for (size_t cache_index = 0; cache_index < layer_caches.size();
+         ++cache_index) {
+      const RegisteredCache& registered_cache = layer_caches[cache_index];
+      const RegisteredCache& stage_cache = layer_staging_caches[cache_index];
+      const int64_t shard_dim =
+          sharded_dimension(registered_cache.role, registered_cache.tensor);
+      CHECK_GE(shard_dim, 0);
+      std::vector<uint64_t> final_ids =
+          is_linear_state_cache(registered_cache.role) ? dst_linear_state_ids
+                                                       : dst_blocks;
+      if (registered_cache.role == KVCacheTensorRole::SSM) {
+        final_ids = expand_checkpoint_ids(final_ids, checkpoint_stride);
+      }
+      if (final_ids.empty()) {
+        continue;
+      }
+
+      // Staging is a compact per-request scratch buffer, not a mirror of
+      // Decode's block allocator. Both sides preserve request block order, so
+      // each shard occupies a contiguous ordinal range.
+      const std::vector<torch::Tensor> shard_tensors =
+          make_staging_shard_views(stage_cache.tensor,
+                                   source_shard_count,
+                                   static_cast<int64_t>(final_ids.size()));
+      const torch::Tensor merged = merge_shard_tensors(registered_cache.role,
+                                                       stage_cache.tensor,
+                                                       layer_staging_caches,
+                                                       shard_tensors,
+                                                       shard_dim,
+                                                       layer_id);
+      copy_merged_cache_rows(registered_cache, final_ids, merged);
+    }
+  }
+  return true;
+}
+
+bool SpecKVCacheTransfer::push_layer_registered_caches_to_staging(
+    const LayerRegisteredCaches& layer_registered_caches,
+    const LayerRegisteredCaches& staging_registered_caches,
+    std::unordered_map<std::string, KVCacheInfo>& merged_kv_infos,
+    std::shared_ptr<NPULayerSynchronizerImpl>& layer_synchronizer,
+    int64_t source_shard_rank,
+    int64_t source_shard_count) {
+  CHECK_GE(source_shard_rank, 0);
+  CHECK_LT(source_shard_rank, source_shard_count);
+  CHECK_EQ(layer_registered_caches.size(), staging_registered_caches.size());
+  std::vector<std::string> keys;
+  keys.reserve(merged_kv_infos.size());
+  for (const auto& pair : merged_kv_infos) {
+    keys.push_back(pair.first);
+  }
+  std::sort(keys.begin(), keys.end());
+
+  Timer total_timer;
+  double layer_wait_seconds = 0.0;
+  double push_seconds = 0.0;
+  bool success = true;
+  for (size_t layer_id = 0; layer_id < layer_registered_caches.size();
+       ++layer_id) {
+    VLOG(5) << "Heterogeneous staged push waiting for layer=" << layer_id
+            << ", source_shard=" << source_shard_rank;
+    Timer layer_wait_timer;
+    layer_synchronizer->synchronize_layer(layer_id);
+    layer_wait_seconds += layer_wait_timer.elapsed_seconds();
+    VLOG(5) << "Heterogeneous staged push layer ready: layer=" << layer_id
+            << ", source_shard=" << source_shard_rank;
+    const auto& layer_caches = layer_registered_caches[layer_id];
+    const auto& layer_staging_caches = staging_registered_caches[layer_id];
+    CHECK_EQ(layer_caches.size(), layer_staging_caches.size());
+
+    for (const std::string& key : keys) {
+      const KVCacheInfo& kv_info = merged_kv_infos.at(key);
+      for (size_t cache_index = 0; cache_index < layer_caches.size();
+           ++cache_index) {
+        const RegisteredCache& source_cache = layer_caches[cache_index];
+        const RegisteredCache& stage_cache = layer_staging_caches[cache_index];
+        // Decode restores CONV/SSM with a synchronous PULL because consuming
+        // their pre-pushed staging rows is not correct on the heterogeneous
+        // Qwen3.5 path. Avoid sending the same large recurrent state twice;
+        // heterogeneous staging PUSH is only useful for target KEY/VALUE.
+        if (is_linear_state_cache(source_cache.role)) {
+          continue;
+        }
+        const std::vector<uint64_t>& src_ids = kv_info.src_blocks;
+        std::vector<uint64_t> dst_ids = kv_info.dst_blocks;
+        if (src_ids.empty() || dst_ids.empty()) {
+          continue;
+        }
+        CHECK_EQ(stage_cache.tensor.size(0) % source_shard_count, 0);
+        const int64_t rows_per_shard =
+            stage_cache.tensor.size(0) / source_shard_count;
+        CHECK_LE(static_cast<int64_t>(dst_ids.size()), rows_per_shard);
+        // Use compact request-local staging rows.  dst_ids are Decode's real
+        // allocator ids and may exceed the bounded staging capacity.
+        for (size_t ordinal = 0; ordinal < dst_ids.size(); ++ordinal) {
+          dst_ids[ordinal] = static_cast<uint64_t>(
+              source_shard_rank * rows_per_shard + ordinal);
+        }
+        CacheIndex destination{kv_info.dst_cluster_id,
+                               stage_cache.cache.cache_id};
+        KvCacheExtParam ext_param{};
+        ext_param.src_layer_range = {0, 0};
+        ext_param.dst_layer_range = {0, 0};
+        ext_param.tensor_num_per_layer = 1;
+        VLOG(5) << "Heterogeneous staged push begin: layer=" << layer_id
+                << ", role=" << source_cache.role.to_string()
+                << ", source_shard=" << source_shard_rank
+                << ", source_cache_id=" << source_cache.cache.cache_id
+                << ", destination_cache_id=" << stage_cache.cache.cache_id;
+        Timer push_timer;
+        const auto ret = llm_data_dist_->PushKvBlocks(
+            source_cache.cache, destination, src_ids, dst_ids, ext_param);
+        push_seconds += push_timer.elapsed_seconds();
+        VLOG(5) << "Heterogeneous staged push end: layer=" << layer_id
+                << ", role=" << source_cache.role.to_string()
+                << ", source_shard=" << source_shard_rank
+                << ", ret=" << std::hex << ret;
+        if (ret != LLM_SUCCESS) {
+          LOG(ERROR) << "Heterogeneous staged PushKvBlocks failed, layer="
+                     << layer_id << ", role=" << source_cache.role.to_string()
+                     << ", source_shard=" << source_shard_rank
+                     << ", destination_cache_id=" << stage_cache.cache.cache_id
+                     << ", ret=" << std::hex << ret;
+          success = false;
+        }
+      }
+    }
+  }
+  LOG(INFO) << "[PD-PERF] Heterogeneous staging push source_shard="
+            << source_shard_rank << ", request_count=" << keys.size()
+            << ", layer_wait_ms=" << layer_wait_seconds * 1000.0
+            << ", push_ms=" << push_seconds * 1000.0
+            << ", total_ms=" << total_timer.elapsed_seconds() * 1000.0;
+  return success;
+}
+
+void SpecKVCacheTransfer::register_hetero_staging_caches(
+    const LayerRegisteredCaches& source_registered_caches,
+    LayerRegisteredCaches& staging_registered_caches,
+    int64_t source_shard_count,
+    bool source_is_sharded) {
+  CHECK_EQ(source_shard_count, kSupportedHeterogeneousSourceShardCount)
+      << "Only Prefill TP2 to Decode TP1 staging is supported.";
+  const int64_t block_size = KVCacheConfig::get_instance().block_size();
+  const int64_t max_tokens_per_batch =
+      SchedulerConfig::get_instance().max_tokens_per_batch();
+  const int64_t max_seqs_per_batch =
+      SchedulerConfig::get_instance().max_seqs_per_batch();
+  CHECK_GT(block_size, 0);
+  CHECK_GT(max_tokens_per_batch, 0);
+  CHECK_GT(max_seqs_per_batch, 0);
+  // Each scheduled sequence may contribute a partially filled final block.
+  const int64_t kv_rows_per_shard =
+      (max_tokens_per_batch + block_size - 1) / block_size + max_seqs_per_batch;
+  staging_registered_caches.clear();
+  staging_registered_caches.resize(source_registered_caches.size());
+  for (size_t layer_id = 0; layer_id < source_registered_caches.size();
+       ++layer_id) {
+    const int64_t checkpoint_stride =
+        get_checkpoint_stride(source_registered_caches[layer_id]);
+    for (const RegisteredCache& source_cache :
+         source_registered_caches[layer_id]) {
+      std::vector<int64_t> shape = source_cache.tensor.sizes().vec();
+      const int64_t shard_dim =
+          sharded_dimension(source_cache.role, source_cache.tensor);
+      CHECK_GE(shard_dim, 0);
+      if (!source_is_sharded) {
+        if (shape[shard_dim] % source_shard_count == 0) {
+          shape[shard_dim] /= source_shard_count;
+        } else {
+          // Homogeneous Decode tensors may already be a single local shard.
+          // Such staging is registered for cache-id symmetry but is unused by
+          // the homogeneous transfer path.
+          CHECK_EQ(shape[shard_dim], 1)
+              << "Unsupported Decode staging shard shape.";
+        }
+      }
+      if (source_cache.role == KVCacheTensorRole::KEY ||
+          source_cache.role == KVCacheTensorRole::VALUE) {
+        shape[0] = std::min(shape[0], kv_rows_per_shard) * source_shard_count;
+      } else if (source_cache.role == KVCacheTensorRole::SSM) {
+        shape[0] = std::min(shape[0], checkpoint_stride) * source_shard_count;
+      } else {
+        shape[0] = source_shard_count;
+      }
+      torch::Tensor stage_tensor =
+          make_page_aligned_staging_tensor(shape,
+                                           source_cache.tensor.options(),
+                                           source_cache.tensor.element_size());
+      staging_registered_caches[layer_id].push_back(
+          register_cache_tensor(layer_id,
+                                KVCacheTensor{source_cache.role,
+                                              stage_tensor,
+                                              source_cache.group_id,
+                                              source_cache.sequence_scoped}));
+    }
+  }
 }
 
 bool SpecKVCacheTransfer::pull_replicated_spec_kv_blocks(
@@ -232,6 +838,10 @@ bool SpecKVCacheTransfer::pull_hetero_kv_blocks(
     const std::vector<uint64_t>& src_linear_state_ids,
     const std::vector<uint64_t>& dst_linear_state_ids) {
   (void)src_addrs;
+  // All heterogeneous cache roles share staging tensors. Serialize the full
+  // Linear State -> Target KV -> Draft KV restore transaction so another
+  // FirstGeneration request cannot overwrite staging between phases.
+  std::lock_guard<std::mutex> restore_lock(hetero_restore_mutex_);
   Timer phase_timer;
   // DataDist PUSH does not expose a Decode-side completion primitive for the
   // large recurrent state tensors. Keep CONV/SSM on the established
@@ -254,7 +864,7 @@ bool SpecKVCacheTransfer::pull_hetero_kv_blocks(
                                       hetero_staging_registered_caches_,
                                       dst_blocks,
                                       /*dst_linear_state_ids=*/{},
-                                      /*source_shard_count=*/2);
+                                      kSupportedHeterogeneousSourceShardCount);
   if (!target_success) {
     return false;
   }
@@ -280,7 +890,8 @@ bool SpecKVCacheTransfer::pull_hetero_kv_blocks(
     const double draft_seconds = phase_timer.elapsed_seconds();
     LOG(INFO) << "Merged heterogeneous TP KV cache (target KV pre-pushed, "
                  "linear state and draft pulled): source_shards="
-              << 2 << ", blocks=" << dst_blocks.size()
+              << kSupportedHeterogeneousSourceShardCount
+              << ", blocks=" << dst_blocks.size()
               << ", linear_states=" << dst_linear_state_ids.size()
               << ", linear_ms=" << linear_seconds * 1000.0
               << ", target_merge_ms=" << target_merge_seconds * 1000.0
